数据库可以使用Cassandra了，Canssandra上的新建用户已完成

+++ Dao/Cache/Connection.go
@@ -6,9 +6,9 @@ import (
 )
 
 type config struct {
-	ExistTime     int    `yaml:"ExistTime"`     //设置每个缓存的存活时间，单位秒
-	RandExistTime int    `yaml:"RandExistTime"` //设置缓存的随机变化范围，单位秒。防止缓存雪崩
-	InitialCache  int    `yaml:"InitialCache"`  //设置初始缓存数量，在服务器启动时写入多少记录，单位条
+	ExistTime     uint64 `yaml:"ExistTime"`     //设置每个缓存的存活时间，单位秒
+	RandExistTime int64  `yaml:"RandExistTime"` //设置缓存的随机变化范围，单位秒。防止缓存雪崩
+	InitialCache  uint64 `yaml:"InitialCache"`  //设置初始缓存数量，在服务器启动时写入多少记录，单位条
 	Network       string `yaml:"Network"`
 	Address       string `yaml:"Address"`
 }
